Adds support for default policy configuration.
